no more inline stack size control
