+++ RcppParallel.Rproj
@@ -1,4 +1,5 @@
 Version: 1.0
+ProjectId: f1ae2e5b-eb1d-4d93-9ae1-29e364a5eaa6
 
 RestoreWorkspace: No
 SaveWorkspace: No
